+++ crates/graph/src/graph_default.rs
@@ -1,0 +1,7 @@
+use crate::graph::Graph;
+
+impl Default for Graph {
+    fn default() -> Self {
+        Self::new()
+    }
+}
