+++ crates/graph/src/graph_edge.rs
@@ -1,15 +1,20 @@
 use super::graph_data::GraphData;
 use super::graph_impl::GraphImpl;
 
-#[allow(dead_code)]
-pub(crate) struct GraphEdge<'a, Data: GraphData> {
-    pub(super) graph: &'a GraphImpl<Data>,
-    pub(super) index: i64,
+pub struct GraphEdge<'a, Data>
+where
+    Data: GraphData,
+{
+    #[allow(dead_code)]
+    pub(crate) graph: &'a GraphImpl<Data>,
+    pub(crate) index: i64,
 }
 
-#[allow(dead_code)]
-impl<'a, Data: GraphData> GraphEdge<'a, Data> {
-    pub(crate) fn index(&self) -> i64 {
+impl<'a, Data> GraphEdge<'a, Data>
+where
+    Data: GraphData,
+{
+    pub fn index(&self) -> i64 {
         self.index
     }
 }
