+++ crates/graph/src/lib.rs
@@ -1,11 +1,19 @@
 mod graph;
 mod graph_data;
 mod graph_data_memory;
+mod graph_data_memory_default;
 mod graph_data_storage;
 mod graph_data_storage_indexes;
+mod graph_data_storage_indexes_serialize;
+mod graph_data_storage_try_from;
+mod graph_default;
 mod graph_edge;
 mod graph_edge_iterator;
 mod graph_impl;
 mod graph_node;
 mod graph_node_iterator;
 mod storage_graph;
+mod storage_graph_try_from;
+
+pub use graph::Graph;
+pub use storage_graph::StorageGraph;
