+++ crates/graph/tests/graph_test.rs
@@ -1,0 +1,6 @@
+use agdb_graph::Graph;
+
+#[test]
+fn derived_from_default() {
+    let _graph = Graph::default();
+}
